Factor shared medical model logic into reusable service module
