+++ medical_ui/catalog/services.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import importlib.util
 import uuid
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass, field
@@ -10,6 +9,8 @@ from typing import Any
 from urllib.error import HTTPError, URLError
 
 from django.conf import settings
+
+from services.medical_models import generate_manifest
 
 BASE_OUTPUT_ROOT = settings.BASE_DIR / "model_downloads"
 BASE_OUTPUT_ROOT.mkdir(parents=True, exist_ok=True)
@@ -26,16 +27,6 @@ class DiscoveryJob:
 JOBS: dict[str, DiscoveryJob] = {}
 JOBS_LOCK = Lock()
 EXECUTOR = ThreadPoolExecutor(max_workers=2, thread_name_prefix="medical-discovery")
-
-
-def _load_medical_script() -> Any:
-    script_path = settings.BASE_DIR / "Docker" / "download_medical_models.py"
-    spec = importlib.util.spec_from_file_location("medical_downloader", script_path)
-    if spec is None or spec.loader is None:
-        raise RuntimeError(f"Unable to load {script_path}")
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    return module
 
 
 def sanitize_output_dir(user_value: str) -> Path:
@@ -64,35 +55,14 @@ def _run_discovery(job_id: str, *, keywords: list[str], limit: int, output_dir: 
     job.status = "running"
 
     try:
-        module = _load_medical_script()
-        output_dir.mkdir(parents=True, exist_ok=True)
-        discovered = module.discover_models(keywords, limit, token=token)
-
-        manifest: list[dict[str, Any]] = []
-        for model_id, model in sorted(discovered.items()):
-            tags = model.get("tags", []) if isinstance(model.get("tags", []), list) else []
-            if not module.is_medical_model(model_id, tags, keywords):
-                continue
-
-            gguf_files = module.extract_gguf_files(model)
-            if not gguf_files:
-                continue
-
-            selected_files = gguf_files if all_files else module.pick_preferred_file(gguf_files)
-
-            if download:
-                for file_name in selected_files:
-                    module.download_file(model_id, file_name, output_dir, token=token)
-
-            manifest.append(
-                {
-                    "id": model_id,
-                    "downloads": model.get("downloads", 0),
-                    "likes": model.get("likes", 0),
-                    "selected_files": selected_files,
-                }
-            )
-
+        manifest = generate_manifest(
+            keywords=keywords,
+            limit=limit,
+            token=token,
+            all_files=all_files,
+            download=download,
+            output_dir=output_dir,
+        )
         job.manifest = manifest
         job.status = "done"
     except (HTTPError, URLError, ValueError, OSError, RuntimeError) as exc:
